# openapi.yml
openapi: "3.0.2"
info:
  title: "Steam Games and Reviews"
  version: "1.0"
servers:
- url: "http://localhost/"
paths:
...
  /reviews:
    post:
      tags:
        - Reviews
      summary: "Create review"
      operationId: add_review
      security:
        - AuthUser:
            - operations_reviews
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Review'
      responses:
        "200":
          description: Review added
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  description:
                    type: string
        "400":
          description: Review already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  description:
                    type: string        
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  description:
                    type: string
    get:
      tags:
        - Reviews
      summary: "Get reviews list"
      operationId: list_reviews
      parameters:
        - name: maxResults
          in: query
          required: false
          schema:
            type: integer
            format: int64      
      security:
        - AuthUser:
            - operations_reviews
      responses:
        "200":
          description: Reviews list returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
        "404":
          description: Review not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  description:
                    type: string
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  description:
                    type: string
  /reviews/{review_id}:
    get:
      tags:
        - Reviews
      summary: "Get review"
      operationId: get_review
      parameters:
        - name: review_id
          in: path
          required: true
          schema:
            type: integer
            format: int64      
      security:
        - AuthUser:
            - operations_reviews
      responses:
        "200":
          description: Review returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
        "404":
          description: Review not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  description:
                    type: string
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  description:
                    type: string       
    put:
      tags:
        - Reviews
      summary: "Update review"
      operationId: update_review
      parameters:
        - name: review_id
          in: path
          required: true
          schema:
            type: integer
            format: int64 
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Review'      
      security:
        - AuthUser:
            - operations_reviews
      responses:
        "200":
          description: Review updated successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
        "404":
          description: Review not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  description:
                    type: string
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  description:
                    type: string
    delete:
      tags:
        - Reviews
      summary: "Delete review"
      operationId: delete_review
      parameters:
        - name: review_id
          in: path
          required: true
          schema:
            type: integer
            format: int64      
      security:
        - AuthUser:
            - operations_reviews
      responses:
        "204":
          description: Review deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  description:
                    type: string
        "404":
          description: Review not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  description:
                    type: string
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  description:
                    type: string                   

  ######WISHLIST######
...
security:
  - app_id: []
...
components:
...
  securitySchemes:
    AuthUser:
      type: oauth2
      description: This API uses OAuth 2 for users
      flows:
        implicit:
          authorizationUrl: https://localhost/
          scopes:
            operations_library: Library (add/delete)
            operations_wishlist: Wishlist (add/delelte)
            operations_reviews: Wishlist (create/edit/delelte)
            operations_user: Edit user info / Logout
    AuthAdmin:
      type: oauth2
      description: This API uses OAuth 2 for admins
      flows:
        implicit:
          authorizationUrl: https://localhost/
          scopes:
            operations_admin: Admin Operations
    app_id:
      type: apiKey
      description: "API key to authorize requests."
      name: appid
      in: query
...
  schemas:
    Review:
      title: Review
      type: object
      properties:
        review_id:
          type: number
          description: The unique id of the recommendation
          example: 85184605
        app_id:
          type: number
          description: The unique id of the app (game)
          example: 292030
        app_name:
          type: string
          description: The name of the app (game)
          example: 'The Witcher 3: Wild Hunt'
        language:
          type: string
          description: Language the user indicated when authoring the review
          example: 'english'
        review:
          type: string
          description: Text of written review
          example: 'One of the best RPGs of all time, worthy of any collection'
        timestamp_created:
          type: number
          description: Date the review was created (unix timestamp)
          example: 1611379970
        timestamp_updated:
          type: number
          description: Date the review was last updated (unix timestamp)
          example: 1611379970
        recommended:
          type: boolean
          description: TRUE if the user recommends the game
          example: true
        votes_helpful:
          type: number
          description: The number of users that found this review helpful
          example: 0
        votes_funny:
          type: number
          description: The number of users that found this review funny
          example: 0
        weighted_vote_score:
          type: number
          description: Helpfulness score
          example: 0.0
        comment_count:
          type: number
          description: Number of comments posted on this review
          example: 0
        steam_purchase:
          type: boolean
          description: TRUE if the user purchased the game on Steam
          example: true
        received_for_free:
          type: boolean
          description: TRUE if the user checked a box saying they got the app for free
          example: false
        written_during_early_access:
          type: boolean
          description: TRUE if the user posted this review while the game was in Early Access
          example: false
        author_steamid:
          type: number
          description: The user’s SteamID
          example: 76561199054755373
        author_num_games_owned:
          type: number
          description: Number of games owned by the user
          example: 5
        author_num_reviews:
          type: number
          description: Number of reviews written by the user
          example: 3
        author_playtime_forever:
          type: number
          description: Lifetime playtime tracked
          example: 5587.0
        author_playtime_last_two_weeks:
          type: number
          description: Playtime tracked in the past two weeks
          example: 3200.0
        author_playtime_at_review:
          type: number
          description: Playtime when the review was written
          example: 5524.0
        author_last_played:
          type: number
          description: Time for when the user last played
          example: 1611383744.0
...